requirements.txt
asgiref==3.10.0
behave==1.3.3
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==8.0.0
Django==5.2.7
factory_boy==3.3.3
Faker==37.12.0
parse==1.20.2
parse_type==0.6.6
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2

djangorestframework~=3.16.1
drf-yasg~=1.21.11